2.9.2 Release
• Fixes an issue where enabling "Deliver Quietly" for an app in iOS 12 would cause notifications to stop being delivered entirely.
