+++ iOS_SDK/OneSignalSDK/Framework/OneSignal.framework/Modules
@@ -1,1 +1,1 @@
-Versions/Current/Modules
+Versions/A/Modules
