+++ iOS_SDK/OneSignalSDK/Framework/OneSignal.framework/OneSignal
@@ -1,1 +1,1 @@
-Versions/Current/OneSignal
+Versions/A/OneSignal
