+++ iOS_SDK/OneSignalSDK/Framework/OneSignal.framework/Resources
@@ -1,0 +1,1 @@
+Versions/A/Resources
