+++ iOS_SDK/OneSignalSDK/Framework/OneSignal.framework/Versions/A/Modules/module.modulemap
@@ -1,5 +1,6 @@
 framework module OneSignal {
-    umbrella header "OneSignal.h"
-    export *
-    module * { export * }
+  umbrella header "OneSignal.h"
+
+  export *
+  module * { export * }
 }
