+++ iOS_SDK/OneSignalSDK/Source/OneSignal.m
@@ -122,7 +122,7 @@ NSString* const kOSSettingsKeyProvidesAppNotificationSettings = @"kOSSettingsKey
 
 @implementation OneSignal
 
-NSString* const ONESIGNAL_VERSION = @"020901";
+NSString* const ONESIGNAL_VERSION = @"020902";
 static NSString* mSDKType = @"native";
 static BOOL coldStartFromTapOnNotification = NO;
